+++ lib/figure_styles.py
@@ -184,9 +184,9 @@ def create_watermark(fig, imagePath ='./lib/NIWA_CMYK_Hor.png',
                      label = 'Climate Present and Past', ax=None, alpha=0.5, loc = "upper left"):
     img = Image.open(imagePath)
     width, height = ax.figure.get_size_inches()*fig.dpi
-    wm_width = np.int(width/4) # make the watermark 1/4 of the figure size
-    scaling = (wm_width / np.float(img.size[0]))
-    wm_height = np.int(float(img.size[1])*float(scaling))
+    wm_width = np.int64(width/4) # make the watermark 1/4 of the figure size
+    scaling = (wm_width / np.float64(img.size[0]))
+    wm_height = np.int64(float(img.size[1])*float(scaling))
     img = img.resize((wm_width, wm_height), Image.ANTIALIAS)
     print(label)
     if label is None:
